Switch to using float buffers to avoid accumulating quantization noise

+++ src/wavewriter.h
@@ -5,7 +5,7 @@ class WaveBuffer
 {
 public:
     int length;
-    short* data;
+    float* data;
 
     WaveBuffer();
     ~WaveBuffer();
